CREATED METHOD TO UPDATE USER

+++ src/main/java/com/squad8/s8orangebackend/service/exceptions/ResourceNotFoundException.java
@@ -2,7 +2,7 @@ package com.squad8.s8orangebackend.service.exceptions;
 
 public class ResourceNotFoundException extends RuntimeException{
 
-    public ResourceNotFoundException(Integer id) {
+    public ResourceNotFoundException(Long id) {
         super("ResourceNotFound. Id: " + id);
     }
 }
